Cleanup optimum/neuron/distributed
